requirements.txt
blinker==1.8.2
cached-property==2.0.1
circuitbreaker==2.0.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
ordered-set==4.1.0
packaging==24.1
Pygments==2.18.0
rich==13.9.4
SQLAlchemy==2.0.36
sqlalchemy-orm==1.2.10
typing-inspect==0.9.0
typing_extensions==4.12.2
Werkzeug==3.1.0
wrapt==1.16.0
zipp==3.20.2
